merges CurationConcerns::Curatable into CurationConcerns::GenericWorkBehavior and adds CurationConcerns::Permissions to CurationConcerns::GenericWorkBehavior
